all: add -dump option
