+++ internal/models/models.go
@@ -31,14 +31,14 @@ type Redir struct {
 // RedirIndex is an extension to Redir, which offers more statistic
 // information such as PV/UV.
 type RedirIndex struct {
-	ID        string    `json:"-"          bson:"_id"`
-	Alias     string    `json:"alias"      bson:"alias"`
-	Kind      AliasKind `json:"kind"       bson:"kind"`
-	URL       string    `json:"url"        bson:"url"`
-	Private   bool      `json:"private"    bson:"private"`
-	ValidFrom time.Time `json:"valid_from" bson:"valid_from"`
-	UV        int64     `json:"uv" bson:"uv"`
-	PV        int64     `json:"pv" bson:"pv"`
+	ID        string    `json:"-"          yaml:"-"          bson:"_id"`
+	Alias     string    `json:"alias"      yaml:"alias"      bson:"alias"`
+	Kind      AliasKind `json:"kind"       yaml:"kind"       bson:"kind"`
+	URL       string    `json:"url"        yaml:"url"        bson:"url"`
+	Private   bool      `json:"private"    yaml:"private"    bson:"private"`
+	ValidFrom time.Time `json:"valid_from" yaml:"valid_from" bson:"valid_from"`
+	UV        int64     `json:"uv"         yaml:"uv"         bson:"uv"`
+	PV        int64     `json:"pv"         yaml:"pv"         bson:"pv"`
 }
 
 // Visit indicates an record of visit pattern.
