+++ short.go
@@ -13,7 +13,6 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"os"
 	"path/filepath"
 	"regexp"
 	"strconv"
@@ -23,7 +22,6 @@ import (
 	"changkun.de/x/redir/internal/db"
 	"changkun.de/x/redir/internal/models"
 	"changkun.de/x/redir/internal/utils"
-	"gopkg.in/yaml.v3"
 )
 
 // op is a short link operator
@@ -46,91 +44,6 @@ func (o op) valid() bool {
 		return true
 	default:
 		return false
-	}
-}
-
-func importFile(fname string) {
-	b, err := os.ReadFile(fname)
-	if err != nil {
-		log.Fatalf("cannot read import file: %v\n", err)
-	}
-
-	var d struct {
-		Short map[string]struct {
-			URL       string `yaml:"url"`
-			Private   bool   `yaml:"private"`
-			ValidFrom string `yaml:"valid_from"`
-		} `yaml:"short"`
-		Random []struct {
-			URL       string `yaml:"url"`
-			Private   bool   `yaml:"private"`
-			ValidFrom string `yaml:"valid_from"`
-		} `yaml:"random"`
-	}
-	err = yaml.Unmarshal(b, &d)
-	if err != nil {
-		log.Fatalf("cannot unmarshal the imported file: %v\n", err)
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*20)
-	defer cancel()
-	for alias, info := range d.Short {
-		var t time.Time
-		if info.ValidFrom != "" {
-			t, err = time.Parse(time.RFC3339, info.ValidFrom)
-			if err != nil {
-				log.Fatalf("incorrect time format, expect RFC3339, but: %v, err: %v", info.ValidFrom, err)
-			}
-		} else {
-			t = time.Now().UTC()
-		}
-
-		r := &models.Redir{
-			Alias:     alias,
-			URL:       info.URL,
-			Kind:      models.KindShort,
-			Private:   info.Private,
-			ValidFrom: t,
-		}
-
-		err = shortCmd(ctx, opUpdate, r)
-		if err != nil {
-			err = shortCmd(ctx, opCreate, r)
-			if err != nil {
-				log.Printf("cannot import alias %v: %v\n", alias, err)
-			}
-		}
-	}
-	for _, info := range d.Random {
-
-		t, err := time.Parse(time.RFC3339, info.ValidFrom)
-		if err != nil {
-			log.Fatalf("incorrect time format, expect RFC3339, but: %v, err: %v", info.ValidFrom, err)
-		}
-		// This might conflict with existing ones, it should be fine
-		// at the moment, the user of redir can always the command twice.
-		if conf.R.Length <= 0 {
-			conf.R.Length = 6
-		}
-		alias := utils.Randstr(conf.R.Length)
-
-		r := &models.Redir{
-			Alias:     alias,
-			URL:       info.URL,
-			Kind:      models.KindRandom,
-			Private:   info.Private,
-			ValidFrom: t,
-		}
-		err = shortCmd(ctx, opUpdate, r)
-		if err != nil {
-			for i := 0; i < 10; i++ { // try 10x maximum
-				err = shortCmd(ctx, opCreate, r)
-				if err != nil {
-					log.Printf("cannot create alias %v: %v\n", alias, err)
-					continue
-				}
-				break
-			}
-		}
 	}
 }
 
@@ -347,6 +260,8 @@ func (s *server) shortHandlerPost(kind models.AliasKind, w http.ResponseWriter, 
 	return
 }
 
+// shortHandlerGet is the core of redir service. It redirects a given
+// alias to the actual destination.
 func (s *server) shortHandlerGet(kind models.AliasKind, w http.ResponseWriter, r *http.Request) (err error) {
 	ctx := r.Context()
 
@@ -398,19 +313,16 @@ func (s *server) shortHandlerGet(kind models.AliasKind, w http.ResponseWriter, r
 		s.cache.Put(alias, red)
 	}
 
-	// redirect the user immediate, but run pv/uv count in background
-	if time.Now().UTC().Sub(red.ValidFrom.UTC()) > 0 {
-		http.Redirect(w, r, red.URL, http.StatusTemporaryRedirect)
-	} else {
-		err = wTmpl.Execute(w, &struct {
+	// Send a wait page if time does not permitting
+	if time.Now().UTC().Sub(red.ValidFrom.UTC()) < 0 {
+		return wTmpl.Execute(w, &struct {
 			ValidFrom string
 			// no timezone, client should conver to local time.
 		}{red.ValidFrom.UTC().Format("2006-01-02T15:04:05")})
-		if err != nil {
-			return
-		}
 	}
 
+	// Finally, let's redirect!
+	http.Redirect(w, r, red.URL, http.StatusTemporaryRedirect)
 	return
 }
 
@@ -589,7 +501,7 @@ func (s *server) sIndex(
 
 	// Serve the index page.
 	w.Header().Add("Content-Type", "text/html")
-	return sTmpl.Execute(w, e)
+	return dTmpl.Execute(w, e)
 }
 
 type indexOutput struct {
